added codenames to each guy
