+++ src/menupages/Parking.tsx
@@ -10,6 +10,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSquareParking } from '@fortawesome/free-solid-svg-icons';
 
 export const parkingPage = {
+    "codename": "parking",
     "header": (<><FontAwesomeIcon icon={faSquareParking} fixedWidth></FontAwesomeIcon> Parking</>),
     "fluidImage": (<></>),
     "body": (<>test</>),
